work around more cgo type weirdness starting with Go 1.10
The compiler seems confused about the SecKeyAlgorithm type. It seems to see this type as a
CFStringRef as it's passed around, but sees it as a *struct___CFString in C function signatures.

Prior to this workaround, the build was failing with this message:

`./certstore_darwin.go:276: cannot use algo (type _Ctype_CFStringRef) as type *_Ctype_struct___CFString in argument to func literal`